correct apply_state size as well

+++ tracer/src/Datadog.Trace/RemoteConfigurationManagement/Protocol/RcmConfigState.cs
@@ -9,7 +9,7 @@ namespace Datadog.Trace.RemoteConfigurationManagement.Protocol
 {
     internal class RcmConfigState
     {
-        public RcmConfigState(string id, long version, string product, uint applyState, string applyError = null)
+        public RcmConfigState(string id, long version, string product, ulong applyState, string applyError = null)
         {
             Id = id;
             Version = version;
@@ -28,7 +28,7 @@ namespace Datadog.Trace.RemoteConfigurationManagement.Protocol
         public string Product { get; }
 
         [JsonProperty("apply_state")]
-        public uint ApplyState { get; }
+        public ulong ApplyState { get; }
 
         [JsonProperty("apply_error")]
         public string ApplyError { get; }
